requirements.txt
gradio
langchain
langchain-docling
sentence-transformers
hopsworks[python] == 4.4.*
transformers
torch
accelerate
python-dotenv
faiss-cpu
numpy
pandas
 

